# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tangula API接口文档\n唐古拉山脉，长江正源，滋润万物。\nTangula 寓意数据的源头，为研发团队提供即取即用的数据服务，帮助团队高效产出。",
        "title": "Tangula",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "zhang.jiejun@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "paths": {
...
        "/resource/platform/tenant": {
            "post": {
                "description": "【登录权限】添加云平台租户。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "平台管理"
                ],
                "summary": "添加云平台租户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "云平台Id",
                        "name": "platformId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问域Id",
                        "name": "domainId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户登录账户",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户登录密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/platform.Tenant"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/resource/platform/tenant/:id": {
            "get": {
                "description": "【登录权限】获取指定租户信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "平台管理"
                ],
                "summary": "获取指定云平台租户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标云平台租户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/platform.Tenant"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "【登录权限】编辑指定云平台租户信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "平台管理"
                ],
                "summary": "编辑指定云平台租户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标云平台租户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问域Id",
                        "name": "domainId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户登录账户",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户登录密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/platform.Tenant"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "【登录权限】删除指定云平台租户。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "平台管理"
                ],
                "summary": "删除指定云平台租户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标云平台租户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
...
        "/resource/platform/{id}/tenant": {
            "get": {
                "description": "【登录权限】获取指定云平台的所有租户信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "平台管理"
                ],
                "summary": "获取云平台所有租户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标云平台ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.Paging"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/platform.Tenant"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Paging": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "totalNum": {
                    "type": "integer"
                }
            }
        },
        "api.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "response": {
                    "type": "object"
                }
            }
        },
...
        "platform.Platform": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "type": "integer"
                },
                "create_user": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passWord": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "platform.Tenant": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "domain_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "passWord": {
                    "type": "string"
                },
                "platform": {
                    "type": "object",
                    "$ref": "#/definitions/platform.Platform"
                },
                "platform_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}